+++ tests/auth.setup.js
@@ -9,7 +9,7 @@ setup('authentication', async ({ page }) => {
 	await page.getByRole('button', { name: 'Log in with username & password' }).click();
 	await page.getByLabel('Email address').fill('admin@fractal.xy');
 	await page.getByLabel('Password').fill('1234');
-	await page.getByRole('button', { name: 'Login' }).click();
+	await page.getByRole('button', { name: 'Log in', exact: true }).click();
 
 	await page.waitForURL('/v2/projects');